+++ client/cody/src/configuration.test.ts
@@ -12,7 +12,7 @@ describe('getConfiguration', () => {
             serverEndpoint: DOTCOM_URL.href,
             codebase: '',
             useContext: 'embeddings',
-            autocomplete: false,
+            autocomplete: true,
             experimentalChatPredictions: false,
             experimentalGuardrails: false,
             experimentalInline: false,
@@ -45,7 +45,7 @@ describe('getConfiguration', () => {
                             'Proxy-Authenticate': 'Basic',
                         }
                     case 'cody.autocomplete.enabled':
-                        return true
+                        return false
                     case 'cody.experimental.chatPredictions':
                         return true
                     case 'cody.experimental.guardrails':
@@ -83,7 +83,7 @@ describe('getConfiguration', () => {
                 'Cache-Control': 'no-cache',
                 'Proxy-Authenticate': 'Basic',
             },
-            autocomplete: true,
+            autocomplete: false,
             experimentalChatPredictions: true,
             experimentalGuardrails: true,
             experimentalInline: true,