+++ client/cody/src/services/DecorationProvider.ts
@@ -8,10 +8,10 @@ const initDecorationType = vscode.window.createTextEditorDecorationType({})
 
 export class DecorationProvider {
     private iconPath: vscode.Uri
-    private fileUri: vscode.Uri | null = null
     private status = CodyTaskState.idle
+    private range = new vscode.Range(0, 0, 0, 0)
+    private editor: vscode.TextEditor | undefined
 
-    private static decorates: DecorationProvider
     private decorations: vscode.DecorationOptions[] = []
     private decorationsForIcon: vscode.DecorationOptions[] = []
 
@@ -24,7 +24,8 @@ export class DecorationProvider {
     private _onDidChange: vscode.EventEmitter<void> = new vscode.EventEmitter<void>()
     public readonly onDidChange: vscode.Event<void> = this._onDidChange.event
 
-    constructor(public id = '', private extPath = '') {
+    constructor(public id: string, private extPath: string, private fileUri: vscode.Uri) {
+        this.editor = vscode.window.visibleTextEditors.find(editor => editor.document.uri.fsPath === fileUri.fsPath)
         // set up icon and register decoration types
         this.iconPath = getIconPath('cody', this.extPath)
         this.decorationTypeIcon = this.makeDecorationType('icon')
@@ -34,76 +35,64 @@ export class DecorationProvider {
             this.decorationTypePending,
             this.decorationTypeError
         )
+        vscode.window.onDidChangeActiveTextEditor(
+            editor => {
+                if (editor?.document.uri.fsPath === this.fileUri.fsPath) {
+                    this.editor = editor
+                    this.decorate()
+                }
+            },
+            null,
+            this._disposables
+        )
     }
-    /**
-     * Getter
-     */
-    public static get instance(): DecorationProvider {
-        return (this.decorates ??= new this())
-    }
+
     /**
      * Highlights line where the codes updated by Cody are located.
      */
-    public async decorate(range: vscode.Range): Promise<void> {
-        if (!this.fileUri || !range) {
+    public decorate(): void {
+        const range = this.range
+        const editor = this.editor
+        if (!editor) {
             return
         }
-        const currentFile = await vscode.workspace.openTextDocument(this.fileUri)
-        if (!currentFile) {
-            return
-        }
-        await vscode.window.showTextDocument(this.fileUri)
+        this.clear()
+        this.decorations.push({ range })
         const rangeStartLine = getSingleLineRange(range.start.line)
         if (this.status === CodyTaskState.error) {
             this.decorationTypePending.dispose()
-            this.decorations.push({ range, hoverMessage: 'Failed Cody Assist #' + this.id })
             this.decorationsForIcon.push({ range: rangeStartLine })
-            vscode.window.activeTextEditor?.setDecorations(this.decorationTypeError, this.decorations)
-            await vscode.window.showTextDocument(this.fileUri, { selection: rangeStartLine })
+            editor.setDecorations(this.decorationTypeError, this.decorations)
             return
         }
         if (this.status === CodyTaskState.fixed) {
             this.decorationTypePending.dispose()
-            this.decorations.push({ range, hoverMessage: 'Cody Assist #' + this.id })
             this.decorationsForIcon.push({ range: rangeStartLine })
-            vscode.window.activeTextEditor?.setDecorations(this.decorationTypeIcon, this.decorationsForIcon)
-            vscode.window.activeTextEditor?.setDecorations(this.decorationTypeDiff, this.decorations)
-            await vscode.window.showTextDocument(this.fileUri, { selection: rangeStartLine })
+            editor.setDecorations(this.decorationTypeIcon, this.decorationsForIcon)
+            editor.setDecorations(this.decorationTypeDiff, this.decorations)
             return
         }
-        vscode.window.activeTextEditor?.setDecorations(this.decorationTypePending, [
+        editor.setDecorations(this.decorationTypePending, [
             { range, hoverMessage: 'Do not make changes to the highlighted code while Cody is working on it.' },
         ])
     }
+
     /**
      * Clear all decorations
      */
-    public async clear(): Promise<void> {
-        if (!this.fileUri) {
-            return
-        }
-        await vscode.workspace.openTextDocument(this.fileUri)
-        this.decorationTypePending.dispose()
-        this.decorationTypeIcon.dispose()
-        this.decorationTypeDiff.dispose()
-        this.decorationTypeError.dispose()
+    public clear(): void {
+        this.decorations = []
+        this.decorationsForIcon = []
     }
-    public setFileUri(uri: vscode.Uri): void {
-        this.fileUri = uri
-    }
+
     /**
      * Define Current States
      */
     public setState(status: CodyTaskState, newRange: vscode.Range): void {
         this.status = status
-        vscode.window.activeTextEditor?.setDecorations(this.decorationTypePending, [newRange])
+        this.range = new vscode.Range(newRange.start.line, 0, Math.max(0, newRange.end.line - 1), 0)
+        this.decorate()
         this._onDidChange.fire()
-    }
-    /**
-     * Remove everything created for task
-     */
-    public remove(): void {
-        this.dispose()
     }
     /**
      * Define styles