+++ client/jetbrains/src/main/java/com/sourcegraph/cody/UpdatableChat.java
@@ -1,6 +1,8 @@
 package com.sourcegraph.cody;
 
 import com.sourcegraph.cody.chat.ChatMessage;
+import com.sourcegraph.cody.context.ContextMessage;
+import java.util.List;
 import org.jetbrains.annotations.NotNull;
 
 public interface UpdatableChat {
@@ -12,9 +14,15 @@ public interface UpdatableChat {
 
   void updateLastMessage(@NotNull ChatMessage message);
 
+  void displayUsedContext(@NotNull List<ContextMessage> contextMessages);
+
   void finishMessageProcessing();
 
   void resetConversation();
+
+  void refreshPanelsVisibility();
+
+  boolean isChatVisible();
 
   void activateChatTab();
 }