+++ client/jetbrains/src/main/java/com/sourcegraph/cody/recipes/SummarizeRecentChangesRecipe.java
@@ -14,7 +14,6 @@ import com.sourcegraph.cody.vcs.Last5ItemsFromCurrentFileFilterOption;
 import com.sourcegraph.cody.vcs.VcsCommitsMetadataLoader;
 import com.sourcegraph.cody.vcs.VcsFilter;
 import com.sourcegraph.cody.vcs.VcsLogFilterOptionsRegistry;
-import java.util.Collections;
 import java.util.Optional;
 import java.util.function.Supplier;
 import org.apache.commons.lang.ArrayUtils;
@@ -71,7 +70,7 @@ public class SummarizeRecentChangesRecipe {
                             chat.activateChatTab();
                             chat.addMessageToChat(
                                 ChatMessage.createHumanMessage(
-                                    "", vcsFilter.getFilterDescription(), Collections.emptyList()));
+                                    "", vcsFilter.getFilterDescription()));
                             chat.addMessageToChat(
                                 ChatMessage.createAssistantMessage("No recent changes found."));
                           });