+++ client/web/src/cody/chat/CodyChatPage.tsx
@@ -105,6 +105,7 @@ export const CodyChatPage: React.FunctionComponent<CodyChatPageProps> = ({
     const codyChatStore = useCodyChat({
         onTranscriptHistoryLoad,
         autoLoadTranscriptFromHistory: false,
+        autoLoadScopeWithRepositories: isSourcegraphApp,
     })
     const {
         initializeNewChat,
@@ -201,7 +202,7 @@ export const CodyChatPage: React.FunctionComponent<CodyChatPageProps> = ({
                     <>
                         Cody answers code questions and writes code for you by leveraging your entire codebase and the
                         code graph.
-                        {!isSourcegraphDotCom && isCTADismissed && (
+                        {!isSourcegraphDotCom && !isSourcegraphApp && isCTADismissed && (
                             <>
                                 {' '}
                                 <Link to="/help/cody#get-cody">Cody is more powerful in the IDE</Link>.