+++ client/web/src/setup-wizard/index.ts
@@ -1,4 +1,4 @@
 export { SetupWizard } from './SetupWizard'
 
-export { SetupStepsContent, SetupStepsRoot } from './components'
+export { SetupStepsContent, SetupStepsRoot, SetupStepsFooter, FooterWidget } from './components'
 export type { StepConfiguration } from './components'