+++ cmd/frontend/codyapp/latest_version.go
@@ -1,0 +1,99 @@
+package codyapp
+
+import (
+	"context"
+	"net/http"
+	"net/url"
+	"time"
+
+	"github.com/sourcegraph/log"
+
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/updatecheck"
+	"github.com/sourcegraph/sourcegraph/internal/conf/deploy"
+)
+
+// RouteCodyAppLatestVersion is the name of the route that that returns a URL where to download the latest Cody App version
+const RouteCodyAppLatestVersion = "codyapp.latest.version"
+
+const gitHubReleaseBaseURL = "https://github.com/sourcegraph/sourcegraph/releases/tag/"
+
+type latestVersion struct {
+	logger           log.Logger
+	manifestResolver updatecheck.UpdateManifestResolver
+}
+
+func (l *latestVersion) Handler() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+		defer cancel()
+		manifest, err := l.manifestResolver.Resolve(ctx)
+		if err != nil {
+			l.logger.Error("failed to resolve manifest", log.Error(err))
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+
+		query := r.URL.Query()
+		target := query.Get("target")
+		arch := query.Get("arch")
+		platform := updatecheck.PlatformString(arch, target) // x86_64-darwin
+
+		releaseURL, err := url.Parse(gitHubReleaseBaseURL)
+		if err != nil {
+			l.logger.Error("failed to create release url from base release url", log.Error(err), log.String("releaseTag", manifest.GitHubReleaseTag()))
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		releaseURL = releaseURL.JoinPath(manifest.GitHubReleaseTag())
+
+		releaseLoc, hasPlatform := manifest.Platforms[platform]
+		// if we have the platform, get it's release URL and redirect to it.
+		// if we don't have it or something goes wrong while converting to a URL, we
+		// redirect to the GitHub release page
+		if hasPlatform {
+			u, err := url.Parse(releaseLoc.URL)
+			if err != nil {
+				l.logger.Error("failed to create release url for platform - redirecting to release page instead",
+					log.Error(err),
+					log.String("platform", platform),
+					log.String("releaseTag", manifest.GitHubReleaseTag()),
+				)
+				http.Redirect(w, r, releaseURL.String(), http.StatusSeeOther)
+				return
+			}
+			releaseURL = u
+		}
+
+		http.Redirect(w, r, releaseURL.String(), http.StatusSeeOther)
+	}
+}
+
+func newLatestVersion(logger log.Logger, resolver updatecheck.UpdateManifestResolver) *latestVersion {
+	return &latestVersion{
+		logger:           logger,
+		manifestResolver: resolver,
+	}
+}
+
+func LatestVersionHandler(logger log.Logger) http.HandlerFunc {
+	var bucket = updatecheck.ManifestBucket
+
+	if deploy.IsDev(deploy.Type()) {
+		bucket = updatecheck.ManifestBucketDev
+	}
+
+	resolver, err := updatecheck.NewGCSManifestResolver(context.Background(), bucket, updatecheck.ManifestName)
+	if err != nil {
+		logger.Error("failed to initialize GCS Manifest resolver",
+			log.String("bucket", bucket),
+			log.String("manifestName", updatecheck.ManifestName),
+			log.Error(err),
+		)
+		return func(w http.ResponseWriter, _ *http.Request) {
+			logger.Warn("GCS Manifest resolver not initialized. Unable to respond with latest App version")
+			w.WriteHeader(http.StatusInternalServerError)
+		}
+	}
+
+	return newLatestVersion(logger, resolver).Handler()
+}